feat: header visible activelink && cleanup
