+++ src/components/AppHeader/AppHeader.tsx
@@ -1,19 +1,28 @@
 import {
   BurgerIcon,
   Button,
+  ListIcon,
   Logo,
+  ProfileIcon,
 } from "@ya.praktikum/react-developer-burger-ui-components";
 import styles from "./appHeader.module.css";
-import { Link } from "react-router-dom";
+import { Link, useLocation } from "react-router-dom";
 import { defaultPath, ordersPath, profilePath } from "../../shared/paths";
 import { useDispatch } from "react-redux";
 import { SET_ACTIVE_NAV_LINK } from "../../services/actions/header";
+import { useEffect, useState } from "react";
 
 const AppHeader = () => {
   const dispatch = useDispatch();
+  const location = useLocation();
   const onClickHandler = (path: string) => {
     dispatch({ type: SET_ACTIVE_NAV_LINK, activeLink: path });
   };
+  const [activeNavLink, setActiveNavLink] = useState("");
+
+  useEffect(() => {
+    setActiveNavLink(location.pathname);
+  }, [location.pathname]);
 
   return (
     <header className={`${styles.header} pt-4 pb-4 pr-5 pl-5`}>
@@ -27,7 +36,9 @@ const AppHeader = () => {
           >
             <BurgerIcon type="primary" />
             <Link
-              className={`${styles.link} ml-2 text text_type_main-default`}
+              className={`${styles.link} ${
+                activeNavLink !== defaultPath && styles.linkDisabled
+              } ml-2 text text_type_main-default`}
               onClick={() => onClickHandler(defaultPath)}
               to={defaultPath}
             >
@@ -40,9 +51,11 @@ const AppHeader = () => {
             type="secondary"
             size="medium"
           >
-            <BurgerIcon type="primary" />
+            <ListIcon type="primary" />
             <Link
-              className={`${styles.link} ${styles.linkDisabled} ml-2 text text_type_main-default`}
+              className={`${styles.link} ${
+                activeNavLink !== ordersPath && styles.linkDisabled
+              } ml-2 text text_type_main-default`}
               onClick={() => onClickHandler(ordersPath)}
               to={ordersPath}
             >
@@ -57,9 +70,11 @@ const AppHeader = () => {
           type="secondary"
           size="medium"
         >
-          <BurgerIcon type="primary" />
+          <ProfileIcon type="primary" />
           <Link
-            className={`${styles.link} ${styles.linkDisabled} ml-2 text text_type_main-default`}
+            className={`${styles.link}   ${
+              activeNavLink !== profilePath && styles.linkDisabled
+            } ml-2 text text_type_main-default`}
             onClick={() => onClickHandler(ordersPath)}
             to={profilePath}
           >
