+++ src/pages/ForgotPasswordPage/ForgotPasswordPage.tsx
@@ -2,11 +2,11 @@ import {
   Button,
   EmailInput,
 } from "@ya.praktikum/react-developer-burger-ui-components";
-import styles from "./common.module.css";
+import styles from "../index.module.css";
 import { Link, useNavigate } from "react-router-dom";
 import { SyntheticEvent, useState } from "react";
-import { checkEmail } from "../services/apis/authorizationApi";
-import { resetPasswordPath } from "../shared/paths";
+import { checkEmail } from "../../services/apis/authorizationApi";
+import { resetPasswordPath } from "../../shared/paths";
 
 const ForgotPasswordPage = () => {
   const navigate = useNavigate();
