+++ src/shared/types/State.ts
@@ -3,6 +3,7 @@ import { Constructor } from "./Constructor";
 import { OrderDetails } from "./OrderDetails";
 import { Ingredients } from "./Ingredients";
 import { UserState } from "../../services/reducers/userReducer";
+import { Header } from "./Header";
 
 export type State = {
   ingredients: Ingredients;
@@ -10,4 +11,5 @@ export type State = {
   orderConstructor: Constructor;
   orderDetails: OrderDetails;
   user: UserState;
+  header: Header;
 };
